workshop8/src/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
requests==2.26.0
streamlit==1.2.0
llama-cpp-python==0.1.77
langchain==0.0.184
pyyaml==6.0
python-dotenv==0.19.1
huggingface_hub==0.16.4